Make secrets endpoint optional (#122)
Our application only retrieves credentials from vault if the endpoint is given, but in our runtime we're enforcing it's ALWAYS given or we crash the app. The variable should either be required everywhere or optional everywhere.

I think optional makes more sense; some instances of reaper may not need to use the credential functionality (like when being run locally).